добавлен модуль conftest.py  и внесены в него фикстуры для тестирования всех функций
