+++ tests/test_processing.py
@@ -2,17 +2,8 @@ import pytest
 from src.processing import filter_by_state, sort_by_date
 
 
-@pytest.fixture
-def sample_data():
-    return [
-        {"id": 1, "state": "EXECUTED"},
-        {"id": 2, "state": "CANCELED"},
-        {"id": 3, "state": "PENDING"},
-        {"id": 4, "state": None},
-        {"id": 5, "state": "EXECUTED"},
-    ]
-
-
+# Тестирование функции `filter_by_state`
+# Параметризация тестов
 @pytest.mark.parametrize(
     "state, expected",
     [
@@ -42,36 +33,8 @@ def test_filter_by_state(sample_data, state, expected):
     assert filter_by_state(sample_data, state) == expected
 
 
-def test_empty_data():
-    assert filter_by_state([], "EXECUTED") == "Ошибка: входная строка пустая."
-
-
-def test_missing_state_key():
-    assert filter_by_state([{"id": 1}], "EXECUTED") == "Ошибка: такого ключа нет."
-
-
-def test_none_state_value():
-    assert (
-        filter_by_state([{"id": 1, "state": None}], "EXECUTED")
-        == "Ошибка: значение 'EXECUTED' не найдено для ключа 'state'."
-    )
-
-
 # Тестирование функции `sort_by_date`
-# Фикстуры для тестирования
-
-
-@pytest.fixture
-def data_sample():
-    return [
-        {"date": "2023-01-01"},
-        {"date": "2022-12-31"},
-        {"date": "2023-01-02"},
-        {"date": "2022-12-30"},
-    ]
-
-
-# Параметризация для тестов
+# Параметризация тестов
 @pytest.mark.parametrize(
     "data, reverse, expected",
     [
@@ -112,19 +75,3 @@ def test_sort_by_date(data, reverse, expected):
     else:
         with pytest.raises(expected):
             sort_by_date(data, reverse)
-
-
-# Дополнительные тесты на ошибки
-def test_empty_list():
-    with pytest.raises(ValueError, match="Ошибка: передан пустой список."):
-        sort_by_date([])
-
-
-def test_missing_date_key():
-    with pytest.raises(ValueError, match="Ошибка: отсутствует ключ 'date'"):
-        sort_by_date([{"not_date": "2023-01-01"}])
-
-
-def test_none_date_value():
-    with pytest.raises(ValueError, match="Ошибка: значение по ключу 'date'"):
-        sort_by_date([{"date": None}])
